Categories is a Tree-structure keeping track of the url sub-directories.
SimplicityHtmlHelper override link(), and adds the url parameter 'lang' to any link produced.

RichTextElementsTable::GetLanguageCodes() returns the language codes used on the page.
RichTextElementsTable::GetLanguagesFor($identifier) returns the language codes the given page are translated into.

RichTextElementsTable is in a state of glory mess right now, as some of it's functionality is transferred into CategoriesTable.