routes updated to handle edit and delete of the pages.
AppController::redirect() overrides the normal redirect() and simply adds the current language as an url parameter.

various updates in EditablePagesController to support the Categories.

CategoriesTable::PathFor() returns the url-path for the given category-id.

CategoriesTable::GetPath() returns the innermost category of the given url-array, or every category.

RichTextElementsTable updated to have a category_id as parent element.

various fixes and cleanup.