Added i18n to table categories, and updated everywhere in order to support several languages in the menus too.
As soon as you save a page in a new language, the entire route to the page in question are also created in the language. The effect is hopefully easy to understand: When you start extending your page in a new language, the menus will contain just the translated pages, and routes to them.